% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rncdb-package.R
\docType{package}
\name{Rncdb-package}
\alias{Rncdb}
\alias{Rncdb-package}
\title{Rncdb: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
